wslog: Use a wider type for line

+++ dumpcap.c
@@ -320,7 +320,7 @@ static gboolean need_timeout_workaround;
 static void
 dumpcap_log_writer(const char *domain, enum ws_log_level level,
                                    struct timespec timestamp,
-                                   const char *file, int line, const char *func,
+                                   const char *file, long line, const char *func,
                                    const char *user_format, va_list user_ap,
                                    void *user_data);
 
@@ -5588,7 +5588,7 @@ main(int argc, char *argv[])
 static void
 dumpcap_log_writer(const char *domain, enum ws_log_level level,
                                    struct timespec timestamp,
-                                   const char *file, int line, const char *func,
+                                   const char *file, long line, const char *func,
                                    const char *user_format, va_list user_ap,
                                    void *user_data _U_)
 {
